+++ frontend/public/changelog.md
@@ -9,6 +9,7 @@
 - Memberikan saran lain ketika pencarian atau detail tidak menemukan hasil
 - Membuat hierarki kata dan menyesuaikan data untuk memfasilitasi itu
 - Menambahkan Google Analytics untuk melacak penggunaan situs
+- Menambahkan batasan pencarian 60 kali/15 menit untuk mencegah penyalahgunaan
 
 ## 2026-02-15 Minggu